Add @Valid annotation to Point record and create associated tests
A new @Valid annotation is introduced and applied to the fields of the Point record. This is done to ensure validity of the point coordinates at the field level. Corresponding test cases have been added to confirm that the annotation is properly applied to the correct components of the Point record.